external_api: Rename module to `api_calls`
